+++ install.py
@@ -1,0 +1,4 @@
+import launch
+
+if not launch.is_installed("nltk"):
+    launch.run_pip("install nltk", "requirements for Unprompted")