+++ lib_unprompted/shared.py
@@ -10,7 +10,7 @@ import sys
 
 class Unprompted:
 	def __init__(self, base_dir="."):
-		self.VERSION = "6.0.0"
+		self.VERSION = "7.0.0"
 
 		print(f"(Unprompted v{self.VERSION} by Therefore Games)")
 		self.log("Initializing Unprompted object...",False,"SETUP")