+++ shortcodes/basic/case.py
@@ -6,14 +6,13 @@ class Shortcode():
 	def run_block(self, pargs, kwargs, context, content):
 		_var = self.Unprompted.shortcode_objects["switch"].switch_var
 
-		if (_var in self.Unprompted.shortcode_user_vars):
-			# Default case
-			if len(pargs) == 0:
-				return(self.Unprompted.parse_alt_tags(content,context))
-			# Matching case
-			elif self.Unprompted.is_equal(self.Unprompted.shortcode_user_vars[_var],self.Unprompted.parse_advanced(pargs[0],context)):
-				self.Unprompted.shortcode_objects["switch"].switch_var = ""
-				return(self.Unprompted.parse_alt_tags(content,context))
+		# Default case
+		if len(pargs) == 0:
+			return(self.Unprompted.parse_alt_tags(content,context))
+		# Matching case
+		elif self.Unprompted.is_equal(_var,self.Unprompted.parse_advanced(pargs[0],context)):
+			self.Unprompted.shortcode_objects["switch"].switch_var = ""
+			return(self.Unprompted.parse_alt_tags(content,context))
 		
 		return("")